add messages.properties

+++ src/main/java/place/reservation/reservations/dto/ProcedureDto.java
@@ -3,11 +3,8 @@ package place.reservation.reservations.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.format.annotation.NumberFormat;
 
 import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
 @Data
@@ -24,5 +21,5 @@ public class ProcedureDto {
     private int durationInMinutes;
 
     @Min(value = 0, message = "Please enter the procedure price")
-    private int price;
+    private double price;
 }
